+++ CHANGELOG.md
@@ -29,6 +29,8 @@ All notable changes to this project will be documented in this file.
 
 - Formatted imports throughout codebase
 
+- Added CQC assessments into both the ratings for data requests and benchmarks datasets.
+
 ### Improved